requirements.txt
# --- Core Dependencies (From your provided list) ---
pandas==2.3.3
numpy==2.3.5
beautifulsoup4==4.14.3
lxml==5.3.0
sentence-transformers==5.2.0
nltk==3.9.2
supabase==2.27.0
python-dotenv==1.2.1
tqdm==4.67.1
torch>=2.5.1
sec-edgar-downloader>=4.0.0
pdfkit>=1.0.0